extract labels to FilterButtonWrap; move LegendButton to ToggleLegend
