--- assets/src/application/ui/layout/entityList/filterBar/buttons/LegendButton/style.scss
@@ -1,12 +1,0 @@
-/** -------------------------------------------------------------------------
- * Event Espresso LegendButton Stylesheet
- * © 2018 Event Espresso
- * ------------------------------------------------------------------------- */
-
-.ee {
-	&-toggle-legend-btn {
-		label {
-			margin-left: 10px;
-		}
-	}
-}
